object: Add support of the raw flag
Object get and head operations can set up `raw` flag in the requests.
When raw flag is set, neofs-node do not try to assemble link object
into original object. Instead they work with it as a usual object.

To check it try to perform object head request without raw flag
and you will see header of the original object:

```
  $ neofs-cli object head --cid <> --oid <> --full-headers
  ...
  Other headers:
    UserHeader: <>
    PublicKey: <>
    Verify: <>
    HomoHash: <>
    PayloadChecksum: <>
    Integrity: <>
```

When you try to use it with raw flag, neofs-node returns header of the
actual link object:

```
  $ neofs-cli object head --cid <> --oid <> --full-headers --raw
  ...
  Other headers:
  UserHeader: <>
    PublicKey: <>
    Verify:< >
    Link: <>
    Link: <>
    Link: <>
    Link: <>
    Link: <>
    PayloadChecksum: <>
    Integrity: <>
```