object: Add copies number support in Put
By default each object is stored in the system according to the rules
of the container. The number of copies optio provides the ability
to override the number of placed copies within a specific object Put.

If CLI user wants to set number of copies to N >= 0, he needs to add
command flag `--copies N`, moreover:

  * object is stored according to container rules if N == 0;

  * otherwise, object is stored in N copies.